Use NPM when publishing
This is really annoying. Generally we want to use `yarn` instead of `npm`.

However, when running `lerna publish`, if it uses `yarn publish` that command will try to prompt the user for a version number, which breaks when running from the lerna script.

See https://github.com/yarnpkg/yarn/issues/2445 for details. Can't really find any other reports of the issue, especially from the `lerna` side of things.