Add simple, incomplete, incorrect autocorrector for stub
Note that instead of calling `double_r?` from our `in_block?` matcher (basically composing them together), I'm not interpolating the string from `double_r?`. This is because we can't extract the matches from the matcher when used in another matcher unless the actual string pattern is present in the matcher we're using.